--- templates/Includes/AssetAdmin_Tools.ss
@@ -1,3 +1,0 @@
-<div id="cms-content-filters-AssetAdmin" class="cms-content-filters">
-	$SearchForm
-</div>
